engine/requirements.txt
watchdog>=4.0
fastapi>=0.115
uvicorn>=0.32
pyyaml>=6.0
python-dotenv>=1.0
networkx>=3.4
gitpython>=3.1
lightrag-hku[api]>=1.4.0,<2.0
google-genai>=1.0
mcp>=1.0
